Initial Code Commit
When I wrote this I wasn't feeling too well, so none of this is actually up to
any standard that I would keep myself. Before I push this up to GitHub, I will
want to refactor it significantly.

This is essentially the first working "quick and dirty" example of attempting to
meet a challenge I issued on Twitter. I intend to embed that tweet in the
README.md. The idea is that a developer randomly chooses three frameworks
they've never used before and combines them in a working demo. It can be
anything, really. Just as long as all three frameworks are actually used.

For the first demo, named "trifoldAlpha", I chose AFNetworking, Magical Record
and UIKit Dynamics. I selected these because I have heard a lot of good things
about the first two and wanted to be ready to use them like a pro in the next
client project instead of sticking with NSURLConnection et al and Core Data
alone (i.e. what I know). I used UIKit Dynamics because I have watched all the
WWDC13 videos about it and have never actually used it.

This commit:

- Sets up the project
- Sets up Cocoapods dependencies
- Writes the .gitignore file (taken from https://github.com/github/gitignore)
- Lays out a **very poor** foundation for the working app

The next step on this will be to thin out the fat view controller into a helper
class or perhaps a UIView and UIViewController subclass, clean up and document
as needed. I'd also like to generalise it a bit and perhaps extend the interface
to allow a movie to be searched for and have the result used in the bounce
simulation. Perhaps even a mini-game.