Slim down the fat view controller (a bit), abstract the API key and add documentation
I don't want to publish code on GitHub that I think sucks (with some exceptions i.e. ZombieArcade), and Q&D projects
often contain lots of sucky code. To that end, I've created a standalone singleton API Client class and reduced the bulk
of the view controller. It's not perfect code but that's not really the point of this concept.

I've also done a bit of searching on the best way to open-source a project that contains an API KEY. After a bit of
playing around, I've added an exclusion in the .gitignore file that prevents MY copy of TFAApiKey.h from entering the
repository. The app builds with the first version of that file that it finds, and there is a copy with dummy keys in the
trifoldAlpha subfolder.

Last, but not least, I've written (and moved) a proper README.md file with explanation and directions.